Move self-signed-cert generation into `CertUtil`.
