+++ src/net-util/package.json
@@ -20,6 +20,7 @@
     "@this/loggy-intf": "*",
     "@this/typey": "*",
     "mime": "^4.0.1",
+    "pem": "^1.14.8",
     "range-parser": "^1.2.1",
     "statuses": "^2.0.1"
   }
